requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.0
email-validator>=2.0.0
